% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialise.R
\name{initialise}
\alias{initialise}
\title{Initialises the optimisation for a jwmodel with all parameters defined}
\usage{
initialise(obj)
}
\arguments{
\item{obj}{jwmodel object to be initialised}
}
\value{
Returns an object of type \code{jwmodel}
}
\description{
Once a jwmodel object has been loaded with all the necessary parameter
information, it must be initialised with this function to ready it for 
optimisation.
}
\examples{
\dontrun{
mymodel <- initialise(mymodel)
}
}
