% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jwmodel.R
\name{jwmodel}
\alias{jwmodel}
\title{Creates a jwmodel object}
\usage{
jwmodel(
  metadata = list(name = "", description = ""),
  judge_types = data.frame(),
  jurisdictions = c(),
  years = c(),
  n_judges = data.frame(),
  judge_departures = data.frame(),
  sitting_days = data.frame(),
  demand = data.frame(),
  judge_progression = data.frame(),
  recruit_limits = data.frame(),
  alloc_limits = data.frame(),
  fixed_costs = data.frame(),
  variable_costs = data.frame(),
  penalty_costs = data.frame(),
  constraints = list(),
  outputs = list(),
  lpmodel = NULL
)
}
\arguments{
\item{metadata}{extendable named list of model metadata, includes model 'name'
and 'description' by default}

\item{judge_types}{data frame containing list of different judge types}

\item{jurisdictions}{list of jurisdictions; default empty}

\item{years}{list of years model will be run for, in accending order; default 
empty}

\item{n_judges}{data frame of initial number of judges by judge type}

\item{judge_departures}{data frame of number of judges leaving the profession
by judge type by year}

\item{sitting_days}{data frame of the sitting day capacity (avg, min, max) for
each WTE judge, by judge type, by year}

\item{demand}{data frame of number of sitting days predicted to be required by 
jurisdiction by year}

\item{judge_progression}{data frame holding the mapping of the proportion of 
judges of one type who are recruited from the pool of another type}

\item{recruit_limits}{data frame holding upper bound constraints on the number 
of judges who can be recruited in any given year}

\item{alloc_limits}{data frame holding the max and min proportions of demand 
which can or must (respectively) be met by a given judge type in a 
given jurisdiction}

\item{fixed_costs}{data frame... default empty}

\item{variable_costs}{data frame... default empty}

\item{penalty_costs}{data frame... default empty}

\item{constraints}{list of constraints (not used)}

\item{outputs}{list of ouputs (not used)}

\item{lpmodel}{default NULL}
}
\value{
Returns an object of type \code{lpmodel}
}
\description{
Recommended process to create and populate a \code{jwmodel} is to use this function
with default values to first create it, then use \code{\link{load_from_file}}
to populate it with parameter information contained in an appropriately 
formatted Excel file.
}
\examples{
# creates new, empty jwmodel object
mymodel <- jwmodel()
}
