% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jwmodel.R
\name{load_from_file}
\alias{load_from_file}
\title{Loads data from a file into a jwmodel object}
\usage{
load_from_file(obj, filepath)
}
\arguments{
\item{obj}{a jwmodel object}

\item{filepath}{a string giving the full filepath name of the Excel file
containing the parameter information to be loaded}
}
\value{
Returns an object of type \code{lpmodel}
}
\description{
Loads data from the specified Excel file into a jwmodel object, priming it
for use. The Excel file must contain the expected worksheets with data in the
expected format.
}
\examples{
\dontrun{
mymodel <- load_from_file(mymodel, "modelparams.xlsx")
}
}
