% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jwmodel.R
\name{optimise}
\alias{optimise}
\title{Runs the optimisation for a jwmodel}
\usage{
optimise(obj)
}
\arguments{
\item{obj}{jwmodel object to be optimised}
}
\description{
Requires the jwmodel to be initialised. Executes the LP optimisation and
rights results into the jwmodel object which is returned.
}
