% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate}
\alias{simulate}
\title{Runs simulation on an optimised jwmodel}
\usage{
simulate(obj, iterations, seed_num)
}
\arguments{
\item{obj}{jwmodel object to be subject to simulation}

\item{iterations}{(optional) integer specifying the total number of demand 
scenarios to be simulated; defaults to 100}

\item{seed_num}{(optional) integer to be used as the seed for the random 
number generator which underpins generation of demand scenarios. Specify if 
you want to re-run a simulation with the same range of demand scenarios.}
}
\value{
Returns an object of type \code{jwmodel}
}
\description{
An optimised jwmodel contains a recommended hiring profile as output. Running 
a simulation means taking this hiring profile (and hence future total judicial 
staff WTE) as fixed, and simulating variations in demand. In each alternate 
demand scenario, allocation of judges to jursidictions is still optimised, but
no alternations to overall headcount are made.
}
\details{
The simulation creates a user-defined number of alternative demand scenarios 
centered around the original expected demand, but varying to account for:
random variation; uncertainty in underlying demand growth; random chances of 
a step-change in demand (e.g. as a result of legislative change).
}
\examples{
\dontrun{
mymodel <- simulate(mymodel, 10000)
}
}
