% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimise.R
\name{optimise}
\alias{optimise}
\title{Runs the optimisation for a jwmodel}
\usage{
optimise(obj, slack_constraints = FALSE)
}
\arguments{
\item{obj}{jwmodel object to be optimised}

\item{slack_constraints}{If TRUE, slack constraints are added for constraints
EQ008 and EQ009. This should prevent most/all infeasibilities. Used for 
diagnosing infeasible models. Default = FALSE.}
}
\value{
Returns an object of type \code{jwmodel}
}
\description{
Requires the jwmodel to be initialised. Executes the LP optimisation and
rights results into the jwmodel object which is returned.
}
\examples{
\dontrun{
mymodel <- optimise(mymodel)
}
}
